Update the create account button when a purchaseToken is already owned
Fixed a bug where an invalid URL in settings are crashing the app
Pre-populated the vote producer screen with the existing votes

Added the logo and blurb to the splash screen

Fetched airdrop tokens customtoken contract

+++ android/app/src/main/java/com/memtrip/eosreach/api/customtokens/Token.kt
@@ -1,7 +1,7 @@
 package com.memtrip.eosreach.api.customtokens
 
 data class Token(
-    val uuid: Long,
+    val uuid: Double,
     val owner: String,
     val customtoken: String,
     val customasset: String
